fix(ci): stop the selftest spec importing the spec file it tests
Playwright refuses to load a test file that another test file imports
("should not import test file"), which failed test collection entirely
(`npx playwright test --list` -> Total: 0 tests in 0 files) once
interactive-state-contrast-selftest.spec.ts started importing
assertContrast/contrastRatio/textContrastFloor straight from
interactive-state-contrast.spec.ts. This took every check in the same
invocation down with it, not just the new ones -- the CI symptom was
gate/full-gate-build's playwright-smoke stage failing exit 1 with no
detail beyond a log path, since the failure happened at collection,
before any test ran.

Moves the pure contrast/threshold math (parseRgb, contrastRatio,
textContrastFloor, assertContrast, EffectiveStyle) into a new non-spec
module, ci/smoke/interactive-contrast.ts, and points both spec files at
it instead of one importing the other. readEffectiveStyle stays in the
browser-driven spec (Locator-only, not needed by the selftest).

Also tags three explanatory comment blocks in
check-interactive-contrast.py (the Greek-label-reveal non-entry, the
home-nav hover fallthrough, and NOT_TEXT_CONTRAST's own rationale) and
one in interactive-state-contrast.spec.ts (TARGETS) with WHY: --
substantive rationale prose with no tag, the same class the prior
commit fixed in two other spots in this same file but missed here.
